Move StringDataNode to the correct namespace
